requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
numpy==2.3.3
openpyxl==3.1.5
pandas==2.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
